reorganized `CareNotfication`
