+++ CareForMe/Core/Controller/View Controllers/TabBar ViewControllers/MainViewController.swift
@@ -121,27 +121,6 @@ extension MainViewController: CareAlertSelectionDelegate {
     
 }
 
-struct CareNotification: Codable, Equatable {
-    let id: UUID
-    let category: String
-    let title: String
-    let text: String
-    let forUserId: String
-    let date: Date
-    
-    static func ==(lhs: CareNotification, rhs: CareNotification) -> Bool {
-        lhs.id == rhs.id
-    }
-    
-    var viewModel: NotificationCellViewModel {
-        NotificationCellViewModel(id: id, category: category, title: title, message: text)
-    }
-    
-    var categoryTitle: String {
-        "\(category): \(title)"
-    }
-}
-
 extension MainViewController: UITableViewDelegate {
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
         CareCollectionView.CareTypeLayout.heightConstant + 16
